pyproject.toml
[tool.poetry]
name = "neurolog"
version = "0.1.0"
description = "Add your description here"
authors = ["0xSolanaceae"]
readme = "README.md"
license = "GPLv3"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = ">=1.6.1"
pandas = ">=2.2.3"
numpy = ">=2.2.2"
python-dateutil = ">=2.9.0.post0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"